+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/manual/log/AbstractLogRecord.java
@@ -101,14 +101,6 @@ public abstract class AbstractLogRecord extends Record {
     @Column(columnName = TAGS, indexOnly = true)
     @SQLDatabase.AdditionalEntity(additionalTables = {ADDITIONAL_TAG_TABLE})
     private List<String> tagsInString;
-    /**
-     * The additional tables need timeBucket for TTL.
-     */
-    @Getter
-    @Setter
-    @Column(columnName = TIME_BUCKET)
-    @SQLDatabase.AdditionalEntity(additionalTables = {ADDITIONAL_TAG_TABLE})
-    private long timeBucket;
 
     @Override
     public String id() {